Swap app and gem
Moves the gem to the root directory and moves the example app to a
directory within the `/spec` directory. This makes it possible to refer
to the gem directly from git or Github.

The `translate` method defined in the controller was causing i18n-tasks
to get confused about the missing keys in the locale file. Renamed it to
`translate_with_resource`.